first reshuffle of functions
