+++ .coveragerc
@@ -4,5 +4,5 @@ branch = False
 [paths]
 tavern =
     tavern/
-    .tox/py38-generic/lib/python3.8/site-packages/tavern/
-    .tox/py38-mqtt/lib/python3.8/site-packages/tavern
+    .tox/py39-generic/lib/python3.9/site-packages/tavern/
+    .tox/py39-mqtt/lib/python3.9/site-packages/tavern
