+++ coverage.sh
@@ -2,9 +2,9 @@
 
 set -ex
 
-tox -c tox-integration.ini -e py38-generic
-tox -c tox-integration.ini -e py38-mqtt
-tox -e py38
+tox -c tox-integration.ini -e py39-generic
+tox -c tox-integration.ini -e py39-mqtt
+tox -e py39 -r
 
 coverage combine --append .coverage tests/integration/.coverage example/mqtt/.coverage
 coverage report -m
