+++ setup.cfg
@@ -49,7 +49,7 @@ exclude =
 console_scripts =
     tavern-ci = tavern.entry:main
 pytest11 =
-    tavern = tavern.testutils.pytesthook
+    tavern = tavern._core.pytest
 
 tavern_http =
     requests = tavern._plugins.rest.tavernhook:TavernRestPlugin
