+++ tavern/_core/dict_util.py
@@ -7,7 +7,8 @@ import string
 from box import Box, box
 import jmespath
 
-from tavern.util.loader import (
+from tavern._core import exceptions
+from tavern._core.loader import (
     ANYTHING,
     ForceIncludeToken,
     RegexSentinel,
@@ -15,7 +16,6 @@ from tavern.util.loader import (
     TypeSentinel,
 )
 
-from . import exceptions
 from .formatted_str import FormattedString
 from .strict_util import StrictSetting, extract_strict_setting
 
@@ -330,8 +330,8 @@ def check_keys_match_recursive(expected_val, actual_val, keys, strict=True):
         True
         >>> check_keys_match_recursive({"a": {"b": "c"}}, {"a": {"b": "d"}}, []) # doctest: +IGNORE_EXCEPTION_DETAIL
         Traceback (most recent call last):
-          File "/home/michael/code/tavern/tavern/tavern/util/dict_util.py", line 223, in check_keys_match_recursive
-        tavern.util.exceptions.KeyMismatchError: Key mismatch: (expected["a"]["b"] = 'c', actual["a"]["b"] = 'd')
+          File "/home/michael/code/tavern/tavern/tavern/_core.util/dict_util.py", line 223, in check_keys_match_recursive
+        tavern._core.util.exceptions.KeyMismatchError: Key mismatch: (expected["a"]["b"] = 'c', actual["a"]["b"] = 'd')
 
     Todo:
         This could be turned into a single-dispatch function for cleaner
