+++ tavern/_core/exceptions.py
