+++ tavern/_core/extfunctions.py
@@ -2,7 +2,8 @@ import functools
 import importlib
 import logging
 
-from . import exceptions
+from tavern._core import exceptions
+
 from .dict_util import deep_dict_merge
 
 
@@ -21,7 +22,7 @@ def get_pykwalify_logger(module):
 
 
 def _getlogger():
-    return get_pykwalify_logger("tavern.util.extfunctions")
+    return get_pykwalify_logger("tavern._core.util.extfunctions")
 
 
 def import_ext_function(entrypoint):
