+++ tavern/_core/testhelpers.py
@@ -1,11 +1,29 @@
 from functools import wraps
 import logging
+import time
 
-from . import exceptions
-from .delay import delay
-from .dict_util import format_keys
+from tavern._core import exceptions
+from tavern._core.dict_util import format_keys
 
 logger = logging.getLogger(__name__)
+
+
+def delay(stage, when, variables):
+    """Look for delay_before/delay_after and sleep
+
+    Args:
+        stage (dict): test stage
+        when (str): 'before' or 'after'
+        variables (dict): Variables to format with
+    """
+
+    try:
+        length = format_keys(stage["delay_{}".format(when)], variables)
+    except KeyError:
+        pass
+    else:
+        logger.debug("Delaying %s request for %.2f seconds", when, length)
+        time.sleep(length)
 
 
 def retry(stage, test_block_config):
