+++ tavern/_plugins/mqtt/client.py
@@ -6,8 +6,8 @@ import time
 
 import paho.mqtt.client as paho
 
-from tavern.util import exceptions
-from tavern.util.dict_util import check_expected_keys
+from tavern._core import exceptions
+from tavern._core.dict_util import check_expected_keys
 
 # MQTT error values
 _err_vals = {
