+++ tavern/_plugins/mqtt/request.py
@@ -4,11 +4,11 @@ import logging
 
 from box import Box
 
-from tavern.request.base import BaseRequest
-from tavern.util import exceptions
-from tavern.util.dict_util import check_expected_keys, format_keys
-from tavern.util.extfunctions import update_from_ext
-from tavern.util.report import attach_yaml
+from tavern._core import exceptions
+from tavern._core.dict_util import check_expected_keys, format_keys
+from tavern._core.extfunctions import update_from_ext
+from tavern._core.report import attach_yaml
+from tavern.request import BaseRequest
 
 logger = logging.getLogger(__name__)
