+++ tavern/_plugins/rest/response.py
@@ -4,11 +4,11 @@ from urllib.parse import parse_qs, urlparse
 
 from requests.status_codes import _codes  # type:ignore
 
-from tavern.response.base import BaseResponse, indent_err_text
-from tavern.testutils.pytesthook.newhooks import call_hook
-from tavern.util import exceptions
-from tavern.util.dict_util import deep_dict_merge
-from tavern.util.report import attach_yaml
+from tavern._core import exceptions
+from tavern._core.dict_util import deep_dict_merge
+from tavern._core.pytest.newhooks import call_hook
+from tavern._core.report import attach_yaml
+from tavern.response import BaseResponse, indent_err_text
 
 logger = logging.getLogger(__name__)
