+++ tavern/helpers.py
@@ -8,10 +8,10 @@ from box import Box
 import jmespath
 import jwt
 
-from tavern.schemas.files import verify_pykwalify
-from tavern.testutils.jmesutils import actual_validation, validate_comparison
-from tavern.util import exceptions
-from tavern.util.dict_util import check_keys_match_recursive, recurse_access_key
+from tavern._core import exceptions
+from tavern._core.dict_util import check_keys_match_recursive, recurse_access_key
+from tavern._core.jmesutils import actual_validation, validate_comparison
+from tavern._core.schema.files import verify_pykwalify
 
 logger = logging.getLogger(__name__)
 
@@ -185,3 +185,31 @@ def validate_content(response, comparisons):
                 )
             except AssertionError as e:
                 raise exceptions.JMESError("Error validating JMES") from e
+
+
+def check_jmespath_match(parsed_response, query, expected=None):
+    """
+    Check that the JMES path given in 'query' is present in the given response
+
+    Args:
+        parsed_response (dict, list): Response list or dict
+        query (str): JMES query
+        expected (str, optional): Possible value to match against. If None,
+            'query' will just check that _something_ is present
+    """
+    actual = jmespath.search(query, parsed_response)
+
+    msg = "JMES path '{}' not found in response".format(query)
+
+    if actual is None:
+        raise exceptions.JMESError(msg)
+
+    if expected is not None:
+        # Reuse dict util helper as it should behave the same
+        check_keys_match_recursive(expected, actual, [], True)
+    elif not actual and not (actual == expected):  # pylint: disable=superfluous-parens
+        # This can return an empty list, but it might be what we expect. if not,
+        # raise an exception
+        raise exceptions.JMESError(msg)
+
+    return actual
