--- tavern/request/__init__.py
@@ -1,3 +1,0 @@
-from .base import BaseRequest
-
-__all__ = ["BaseRequest"]
