--- tavern/response/__init__.py
@@ -1,3 +1,0 @@
-from .base import BaseResponse
-
-__all__ = ["BaseResponse"]
