+++ tests/unit/test_helpers.py
@@ -6,17 +6,7 @@ from unittest.mock import Mock, patch
 
 import _pytest
 import pytest
-import yaml
-
-from tavern.core import run
-from tavern.schemas.extensions import validate_file_spec
-from tavern.testutils.helpers import (
-    validate_content,
-    validate_pykwalify,
-    validate_regex,
-)
 from tavern.testutils.pytesthook.item import YamlItem
-from tavern.util import exceptions
 from tavern.util.dict_util import _check_and_format_values, format_keys
 from tavern.util.loader import ForceIncludeToken
 from tavern.util.strict_util import (
@@ -24,6 +14,24 @@ from tavern.util.strict_util import (
     StrictSetting,
     validate_and_parse_option,
 )
+import yaml
+
+from tavern._core import exceptions
+from tavern._core.dict_util import _check_and_format_values, format_keys
+from tavern._core.loader import ForceIncludeToken
+from tavern._core.pytest.item import YamlItem
+from tavern._core.schema.extensions import validate_file_spec
+from tavern._core.strict_util import (
+    StrictLevel,
+    validate_and_parse_option,
+)
+from tavern.core import run
+from tavern.helpers import (
+    validate_content,
+    validate_pykwalify,
+    validate_regex,
+)
+from tavern.util import exceptions
 
 
 class FakeResponse:
@@ -143,7 +151,7 @@ class TestTavernRepr:
         """Does not call tavern repr for non tavern errors"""
         fake_info = self._make_fake_exc_info(RuntimeError)
 
-        with patch("tavern.testutils.pytesthook.item.ReprdError") as rmock:
+        with patch("tavern._core.pytesthook.item.ReprdError") as rmock:
             fake_item.repr_failure(fake_info)
 
         assert not rmock.called
@@ -154,7 +162,7 @@ class TestTavernRepr:
         fake_info = self._make_fake_exc_info(exceptions.BadSchemaError)
 
         with patch.object(fake_item.config, "getini", return_value=ini_flag):
-            with patch("tavern.testutils.pytesthook.item.ReprdError") as rmock:
+            with patch("tavern._core.pytesthook.item.ReprdError") as rmock:
                 fake_item.repr_failure(fake_info)
 
         assert not rmock.called
@@ -164,7 +172,7 @@ class TestTavernRepr:
         fake_info = self._make_fake_exc_info(exceptions.InvalidSettingsError)
 
         with patch.object(fake_item.config, "getini", return_value=True):
-            with patch("tavern.testutils.pytesthook.item.ReprdError") as rmock:
+            with patch("tavern._core.pytesthook.item.ReprdError") as rmock:
                 fake_item.repr_failure(fake_info)
 
         assert not rmock.called
@@ -174,7 +182,7 @@ class TestTavernRepr:
         fake_info = self._make_fake_exc_info(exceptions.InvalidSettingsError)
 
         with patch.object(fake_item.config, "getoption", return_value=True):
-            with patch("tavern.testutils.pytesthook.item.ReprdError") as rmock:
+            with patch("tavern._core.pytesthook.item.ReprdError") as rmock:
                 fake_item.repr_failure(fake_info)
 
         assert not rmock.called
@@ -277,7 +285,7 @@ class TestCheckParseValues(object):
         "item", [[134], {"a": 2}, yaml, yaml.load, yaml.SafeLoader]
     )
     def test_warns_bad_type(self, item):
-        with patch("tavern.util.dict_util.logger.warning") as wmock:
+        with patch("tavern._core.util.dict_util.logger.warning") as wmock:
             _check_and_format_values("{fd}", {"fd": item})
 
         assert wmock.called_with(
@@ -288,7 +296,7 @@ class TestCheckParseValues(object):
 
     @pytest.mark.parametrize("item", [1, "a", 1.3, format_keys("{s}", dict(s=2))])
     def test_no_warn_good_type(self, item):
-        with patch("tavern.util.dict_util.logger.warning") as wmock:
+        with patch("tavern._core.util.dict_util.logger.warning") as wmock:
             _check_and_format_values("{fd}", {"fd": item})
 
         assert not wmock.called
