+++ tests/unit/test_request.py
@@ -1,12 +1,15 @@
-import attr
 from contextlib import ExitStack
 import os
-import pytest
-import requests
-from requests.cookies import RequestsCookieJar
 import tempfile
 from unittest.mock import Mock
 
+import attr
+import pytest
+import requests
+from requests.cookies import RequestsCookieJar
+
+from tavern._core import exceptions
+from tavern._core.extfunctions import update_from_ext
 from tavern._plugins.rest.request import (
     RestRequest,
     _check_allow_redirects,
@@ -14,8 +17,6 @@ from tavern._plugins.rest.request import (
     _read_expected_cookies,
     get_request_args,
 )
-from tavern.util import exceptions
-from tavern.util.extfunctions import update_from_ext
 
 
 @pytest.fixture(name="req")
